CS-211 adds support for mulitple bundle assignment strategies, including a maximum bundle assignment strategy (aka "snowpack" style)
